chore: move build_initial_state from main to state_utils
