+++ crates/bin/prove_block/src/main.rs
@@ -20,15 +20,10 @@ use starknet_os::crypto::poseidon::PoseidonHash;
 use starknet_os::error::SnOsError::Runner;
 use starknet_os::execution::helper::{ContractStorageMap, ExecutionHelperWrapper};
 use starknet_os::io::input::StarknetOsInput;
-use starknet_os::starknet::business_logic::fact_state::contract_class_objects::{
-    get_ffc_for_contract_class_facts, ContractClassLeaf,
-};
 use starknet_os::starknet::business_logic::fact_state::contract_state_objects::ContractState;
-use starknet_os::starknet::business_logic::fact_state::state::SharedState;
 use starknet_os::starknet::business_logic::utils::write_class_facts;
 use starknet_os::starknet::starknet_storage::CommitmentInfo;
-use starknet_os::starkware_utils::commitment_tree::base_types::{Height, Length, NodePath, TreeIndex};
-use starknet_os::starkware_utils::commitment_tree::binary_fact_tree::BinaryFactTree;
+use starknet_os::starkware_utils::commitment_tree::base_types::{Height, Length, NodePath};
 use starknet_os::starkware_utils::commitment_tree::patricia_tree::nodes::{BinaryNodeFact, EdgeNodeFact};
 use starknet_os::starkware_utils::commitment_tree::patricia_tree::patricia_tree::PatriciaTree;
 use starknet_os::storage::storage::{Fact, FactFetchingContext};
@@ -40,10 +35,12 @@ use utils::get_subcalled_contracts_from_tx_traces;
 
 use crate::reexecute::format_commitment_facts;
 use crate::rpc_utils::{CachedRpcStorage, PathfinderClassProof};
+use crate::state_utils::build_initial_state;
 use crate::types::starknet_rs_tx_to_internal_tx;
 
 mod reexecute;
 mod rpc_utils;
+mod state_utils;
 mod types;
 mod utils;
 
@@ -64,93 +61,6 @@ fn init_logging() {
         .format_timestamp(None)
         .try_init()
         .expect("Failed to configure env_logger");
-}
-
-// build state representing the end of the previous block on which the current
-// block can be built.
-// inspiration: TestSharedState::apply_state_updates_starknet_api()
-//
-// Returns:
-// * a set of all contracts accessed
-// * a SharedState object representing storage for the changes provided
-async fn build_initial_state(
-    ffc: FactFetchingContext<CachedRpcStorage, PedersenHash>,
-    provider: &JsonRpcClient<HttpTransport>,
-    block_number: u64,
-    address_to_class_hash: HashMap<Felt252, Felt252>,
-    address_to_nonce: HashMap<Felt252, Felt252>,
-    class_hash_to_compiled_class_hash: HashMap<Felt252, Felt252>,
-    storage_updates: HashMap<Felt252, HashMap<Felt252, Felt252>>,
-) -> Result<(HashSet<Felt252>, SharedState<CachedRpcStorage, PedersenHash>), Box<dyn Error>> {
-    let shared_state = SharedState::empty(ffc).await?;
-
-    let accessed_addresses_felts: HashSet<_> =
-        address_to_class_hash.keys().chain(address_to_nonce.keys()).chain(storage_updates.keys()).collect();
-    let accessed_addresses: Vec<TreeIndex> = accessed_addresses_felts.iter().map(|x| x.to_biguint()).collect();
-
-    let mut facts = None;
-    let mut ffc = shared_state.ffc;
-    let mut current_contract_states: HashMap<TreeIndex, ContractState> =
-        shared_state.contract_states.get_leaves(&mut ffc, &accessed_addresses, &mut facts).await?;
-
-    // Update contract storage roots with cached changes.
-    let empty_updates = HashMap::new();
-    let mut updated_contract_states = HashMap::new();
-    for address in accessed_addresses_felts {
-        // unwrap() is safe as an entry is guaranteed to be present with `get_leaves()`.
-        let tree_index = address.to_biguint();
-        let updates = storage_updates.get(address).unwrap_or(&empty_updates);
-        let nonce = address_to_nonce.get(address).cloned();
-        let mut class_hash = address_to_class_hash.get(address).cloned();
-        if class_hash.is_none() {
-            let resp = provider.get_class_hash_at(BlockId::Number(block_number), address).await;
-            class_hash = if let Ok(class_hash) = resp { Some(class_hash) } else { Some(Felt252::ZERO) };
-        }
-        let updated_contract_state =
-            current_contract_states.remove(&tree_index).unwrap().update(&mut ffc, updates, nonce, class_hash).await?;
-
-        updated_contract_states.insert(tree_index, updated_contract_state);
-    }
-
-    // Apply contract changes on global root.
-    log::debug!("Updating contract state tree with {} modifications...", accessed_addresses.len());
-    let global_state_modifications: Vec<_> = updated_contract_states.into_iter().collect();
-
-    let updated_global_contract_root =
-        shared_state.contract_states.update(&mut ffc, global_state_modifications, &mut facts).await?;
-
-    let mut ffc_for_contract_class = get_ffc_for_contract_class_facts(&ffc);
-
-    let updated_contract_classes = match shared_state.contract_classes {
-        Some(tree) => {
-            log::debug!(
-                "Updating contract class tree with {} modifications...",
-                class_hash_to_compiled_class_hash.len()
-            );
-            let modifications: Vec<_> = class_hash_to_compiled_class_hash
-                .into_iter()
-                .map(|(key, value)| (key.to_biguint(), ContractClassLeaf::create(value)))
-                .collect();
-            Some(tree.update(&mut ffc_for_contract_class, modifications, &mut facts).await?)
-        }
-        None => {
-            assert_eq!(class_hash_to_compiled_class_hash.len(), 0, "contract_classes must be concrete before update.");
-            None
-        }
-    };
-
-    let accessed_addresses: HashSet<_> = accessed_addresses.into_iter().map(Felt252::from).collect();
-
-    Ok((
-        accessed_addresses,
-        SharedState {
-            contract_states: updated_global_contract_root,
-            contract_classes: updated_contract_classes,
-            ffc,
-            ffc_for_class_hash: ffc_for_contract_class,
-            contract_addresses: Default::default(),
-        },
-    ))
 }
 
 fn compute_class_commitment(
